feat(admin): Constrain display of more than 3 connected services
Because:
* Our support agents have noted how much time it can take to scroll through a long list of connected services

This commit:
* Displays connected services in a 'details' / 'summary' component if there are more than 3
* Renames AttachedClients to ConnectedServices and moves this component into its own file

Closes FXA-5728
