+++ packages/fxa-admin-panel/src/components/AccountSearch/Account/index.test.tsx
@@ -77,109 +77,7 @@ let accountResponse: AccountProps = {
       enabled: true,
     },
   ],
-  attachedClients: [
-    // The following values are based on a real response with modified ID values.
-    {
-      clientId: null,
-      createdTime: new Date(Date.now() - 60 * 60 * 1e3).getTime(),
-      createdTimeFormatted: '1 hour ago',
-      deviceId: 'xxxxxxxx-did-1',
-      deviceType: 'desktop',
-      lastAccessTime: new Date(Date.now() - 5 * 1e3).getTime(),
-      lastAccessTimeFormatted: '5 seconds ago',
-      location: {
-        city: 'Orlando',
-        country: 'USA',
-        state: 'FL',
-        stateCode: null,
-      },
-      name: "UserX's Nightly on machine-xyz",
-      os: 'Mac OS X',
-      userAgent: 'Chrome 89',
-      sessionTokenId: 'xxxxxxxx-stid-1',
-      refreshTokenId: null,
-    },
-    {
-      clientId: 'xxxxxxxx-clid-1',
-      createdTime: new Date(Date.now() - 60 * 60 * 1e3).getTime(),
-      createdTimeFormatted: '1 hour ago',
-      deviceId: null,
-      deviceType: null,
-      lastAccessTime: new Date(Date.now() - 5 * 1e3).getTime(),
-      lastAccessTimeFormatted: '5 seconds ago',
-      location: {
-        city: null,
-        country: null,
-        state: null,
-        stateCode: null,
-      },
-      name: 'Pocket',
-      os: null,
-      userAgent: '',
-      sessionTokenId: null,
-      refreshTokenId: null,
-    },
-    {
-      clientId: 'xxxxxxxx-mzvpn-1',
-      createdTime: new Date(Date.now() - 60 * 60 * 1e3).getTime(),
-      createdTimeFormatted: '1 hour ago',
-      deviceId: null,
-      deviceType: null,
-      lastAccessTime: new Date(Date.now() - 5 * 1e3).getTime(),
-      lastAccessTimeFormatted: '5 seconds ago',
-      location: {
-        city: null,
-        country: null,
-        state: null,
-        stateCode: null,
-      },
-      name: 'Mozilla VPN',
-      os: null,
-      userAgent: '',
-      sessionTokenId: null,
-      refreshTokenId: 'xxxxxxxx-rtid-1',
-    },
-    {
-      clientId: null,
-      createdTime: new Date(Date.now() - 60 * 60 * 1e3).getTime(),
-      createdTimeFormatted: '1 hour ago',
-      deviceId: null,
-      deviceType: null,
-      lastAccessTime: new Date(Date.now() - 5 * 1e3).getTime(),
-      lastAccessTimeFormatted: '5 seconds ago',
-      location: {
-        city: null,
-        country: null,
-        state: null,
-        stateCode: null,
-      },
-      name: 'Firefox 97, Ubuntu',
-      os: 'Ubuntu',
-      userAgent: 'Firefox 97',
-      sessionTokenId: 'xxxxxxxx-stid-2',
-      refreshTokenId: null,
-    },
-    {
-      clientId: null,
-      createdTime: new Date(Date.now() - 60 * 60 * 1e3).getTime(),
-      createdTimeFormatted: '1 hour ago',
-      deviceId: null,
-      deviceType: null,
-      lastAccessTime: new Date(Date.now() - 5 * 1e3).getTime(),
-      lastAccessTimeFormatted: '5 seconds ago',
-      location: {
-        city: null,
-        country: null,
-        state: null,
-        stateCode: null,
-      },
-      name: 'Mobile Safari 15, iPhone',
-      os: 'iOS',
-      userAgent: 'Mobile Safari 15',
-      sessionTokenId: 'xxxxxxxx-stid-3',
-      refreshTokenId: null,
-    },
-  ],
+  attachedClients: [],
   securityEvents: [],
 };
 
@@ -288,57 +186,6 @@ it('displays the account recovery key status', async () => {
   expect(getByTestId('recovery-keys-created-at')).toBeInTheDocument();
   expect(getByTestId('recovery-keys-verified')).toBeInTheDocument();
   expect(getByTestId('recovery-keys-enabled')).toBeInTheDocument();
-});
-
-it('displays the attached clients', async () => {
-  const { findAllByTestId } = render(
-    <MockedProvider>
-      <Account {...accountResponse} />
-    </MockedProvider>
-  );
-
-  const fields: Record<string, HTMLElement[]> = {
-    clientFields: await findAllByTestId('attached-clients-client'),
-    deviceTypeFields: await findAllByTestId('attached-clients-device-type'),
-    userAgentFields: await findAllByTestId('attached-clients-user-agent'),
-    osFields: await findAllByTestId('attached-clients-os'),
-    createdAtFields: await findAllByTestId('attached-clients-created-at'),
-    lastAccessFields: await findAllByTestId(
-      'attached-clients-last-accessed-at'
-    ),
-    locationFields: await findAllByTestId('attached-clients-location'),
-    clientIdFields: await findAllByTestId('attached-clients-client-id'),
-    deviceIdFields: await findAllByTestId('attached-clients-device-id'),
-    sessionTokenIdFields: await findAllByTestId(
-      'attached-clients-session-token-id'
-    ),
-    refreshTokenIdFields: await findAllByTestId(
-      'attached-clients-refresh-token-id'
-    ),
-  };
-
-  // Ensure that only known data is rendered
-  expect(fields['clientFields']).toHaveLength(5);
-  expect(fields['deviceTypeFields']).toHaveLength(1);
-  expect(fields['userAgentFields']).toHaveLength(3);
-  expect(fields['osFields']).toHaveLength(3);
-  expect(fields['createdAtFields']).toHaveLength(5);
-  expect(fields['lastAccessFields']).toHaveLength(5);
-  expect(fields['locationFields']).toHaveLength(1);
-  expect(fields['clientIdFields']).toHaveLength(2);
-  expect(fields['deviceIdFields']).toHaveLength(1);
-  expect(fields['sessionTokenIdFields']).toHaveLength(3);
-  expect(fields['refreshTokenIdFields']).toHaveLength(1);
-
-  // Make sure fields have some sort of content
-  Object.values(fields)
-    .flatMap((x) => x)
-    .forEach((x) => expect(x).toHaveTextContent(/N\/A|\w{3,}/));
-
-  // Make sure device types are mobile, desktop.
-  fields.deviceTypeFields.forEach((type) =>
-    expect(type).toHaveTextContent(/mobile|desktop|unknown/gi)
-  );
 });
 
 it('displays secondary emails', async () => {
