+++ packages/fxa-admin-panel/src/components/AccountSearch/Account/index.tsx
@@ -10,8 +10,6 @@ import {
   SecurityEvents as SecurityEventsType,
   Totp as TotpType,
   RecoveryKeys as RecoveryKeysType,
-  AttachedClient as AttachedClientType,
-  Location,
   LinkedAccount as LinkedAccountType,
   BounceType,
   BounceSubType,
@@ -21,6 +19,7 @@ import { AdminPanelFeature } from 'fxa-shared/guards';
 import BOUNCE_DESCRIPTIONS from './bounce-descriptions';
 import Guard from '../../Guard';
 import Subscription from '../Subscription';
+import { ConnectedServices } from '../ConnectedServices';
 
 export type AccountProps = AccountType & {
   onCleared: () => void;
@@ -580,25 +579,7 @@ export const Account = ({
           <li className="account-li">
             <h3 className="account-header">Connected Services</h3>
           </li>
-          {attachedClients && attachedClients.length > 0 ? (
-            <>
-              {attachedClients.map((attachedClient: AttachedClientType) => (
-                <AttachedClients
-                  key={`${attachedClient.name}-${
-                    attachedClient.sessionTokenId ||
-                    attachedClient.refreshTokenId ||
-                    attachedClient.clientId ||
-                    'Unknown'
-                  }-${attachedClient.createdTime}`}
-                  {...attachedClient}
-                />
-              ))}
-            </>
-          ) : (
-            <li className="account-li account-border-info">
-              This account has nothing attached.
-            </li>
-          )}
+          <ConnectedServices services={attachedClients} />
         </Guard>
       </ul>
 
@@ -895,88 +876,7 @@ const RecoveryKeys = ({ verifiedAt, createdAt, enabled }: RecoveryKeysType) => {
   );
 };
 
-const AttachedClients = ({
-  clientId,
-  createdTime,
-  createdTimeFormatted,
-  deviceId,
-  deviceType,
-  lastAccessTime,
-  lastAccessTimeFormatted,
-  location,
-  name,
-  os,
-  userAgent,
-  sessionTokenId,
-  refreshTokenId,
-}: AttachedClientType) => {
-  const testId = (id: string) => `attached-clients-${id}`;
-  return (
-    <div className="account-li account-border-info">
-      <table className="pt-1" aria-label="simple table">
-        <tbody>
-          <ResultTableRow
-            label="Client"
-            value={format.client(name, clientId)}
-            testId={testId('client')}
-          />
-          <ResultTableRow
-            label="Device Type"
-            value={deviceType}
-            testId={testId('device-type')}
-          />
-          <ResultTableRow
-            label="User Agent"
-            value={userAgent}
-            testId={testId('user-agent')}
-          />
-          <ResultTableRow
-            label="Operating System"
-            value={os}
-            testId={testId('os')}
-          />
-          <ResultTableRow
-            label="Created At"
-            value={format.time(createdTime, createdTimeFormatted)}
-            testId={testId('created-at')}
-          />
-          <ResultTableRow
-            label="Last Used"
-            value={format.time(lastAccessTime, lastAccessTimeFormatted)}
-            testId={testId('last-accessed-at')}
-          />
-          <ResultTableRow
-            label="Location"
-            value={format.location(location)}
-            testId={testId('location')}
-          />
-          <ResultTableRow
-            label="Client ID"
-            value={clientId || 'N/A'}
-            testId={testId('client-id')}
-          />
-          <ResultTableRow
-            label="Device ID"
-            value={deviceId || 'N/A'}
-            testId={testId('device-id')}
-          />
-          <ResultTableRow
-            label="Session Token ID"
-            value={sessionTokenId || 'N/A'}
-            testId={testId('session-token-id')}
-          />
-          <ResultTableRow
-            label="Refresh Token ID"
-            value={refreshTokenId || 'N/A'}
-            testId={testId('refresh-token-id')}
-          />
-        </tbody>
-      </table>
-    </div>
-  );
-};
-
-const ResultTableRow = ({
+export const ResultTableRow = ({
   label,
   value,
   testId,
@@ -999,50 +899,6 @@ const ResultTableRow = ({
       <td data-testid={testId}>{value}</td>
     </tr>
   );
-};
-
-type Nullable<T> = T | null;
-
-const format = {
-  location(location?: Nullable<Location>) {
-    if (
-      !location ||
-      (!location.city &&
-        !location.state &&
-        !location.stateCode &&
-        !location.country &&
-        !location.countryCode)
-    ) {
-      return null;
-    }
-
-    return (
-      <>
-        {[
-          location.city || <i>Unknown City</i>,
-          location.state || location.stateCode || '',
-          location.country || location.country || <i>Unknown Country</i>,
-        ].join(', ')}
-      </>
-    );
-  },
-  time(raw?: Nullable<number>, formatted?: Nullable<string>) {
-    if (!raw || raw < 1) return null;
-
-    return (
-      <>
-        {dateFormat(new Date(raw), DATE_FORMAT)}
-        {formatted ? <i> ({formatted})</i> : <></>}
-      </>
-    );
-  },
-  client(name?: Nullable<string>, clientId?: Nullable<string>) {
-    return (
-      <>
-        {name} {clientId && <i>[{clientId}]</i>}
-      </>
-    );
-  },
 };
 
 export default Account;
